+++ changelog.rst
@@ -4,6 +4,7 @@ Upcoming (TBD)
 Bug fixes:
 ----------
 * Restore cursor shape behaviour for Emacs mode
+* Suggest columns, not datatypes, after a column literally named ``type`` in a ``SELECT`` list.
 
 Features:
 ---------